Rename mod to token_management
