+++ crates/bitwarden-auth/src/lib.rs
@@ -8,7 +8,7 @@ mod auth_client;
 
 pub mod identity;
 pub mod registration;
-pub mod renew;
+pub mod token_management;
 pub mod send_access;
 
 pub(crate) mod api; // keep internal to crate
